+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ### 2.3.2 (TBD)
 
-- Bug: On a MacOS, files generated under /etc/resolver as the result of using include-suffixes in the cluster config, are now properly removed on quit.
+- Change: The list of subnets that the virtual network interface will route is configured dynamically and will follow changes in the cluster.
+- Change: The client no longer need permission to list nodes.
+- Change: The traffic-manager requires permissions to list nodes, pods, and to create a dummy service (in the manager's namespace).
+- Bugfix: Telepresence will now find the pod CIDRs even if its not listed in the nodes.
+- Bugfix: Telepresence no longer requires permission to create a service in order to determine the service subnet accurately.
+- Bugfix: Subnets covered by other subnets are now removed and thus never added to the virtual network interface.
+- Bugfix: On a MacOS, files generated under /etc/resolver as the result of using include-suffixes in the cluster config, are now properly removed on quit.
 
 ### 2.3.1 (June 14, 2021)